requirements.txt
Django==4.2.5
channels==3.0.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
django-shortuuidfield==0.1.3